+++ components/MealsList/MealsList.js
@@ -1,0 +1,37 @@
+import { View, FlatList, StyleSheet } from "react-native";
+
+import MealItem from "./MealItem";
+
+function MealsList({ items }) {
+  function renderMealItem(itemData) {
+    const item = itemData.item;
+    const mealItemProps = {
+      id: item.id,
+      title: item.title,
+      imageUrl: item.imageUrl,
+      duration: item.duration,
+      complexity: item.complexity,
+      affordablity: item.affordablity,
+    };
+    return <MealItem {...mealItemProps} />;
+  }
+
+  return (
+    <View style={styles.container}>
+      <FlatList
+        data={items}
+        keyExtractor={(item) => item.id}
+        renderItem={renderMealItem}
+      />
+    </View>
+  );
+}
+
+export default MealsList;
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 16,
+  },
+});
