+++ screens/FavoritesScreen.js
@@ -1,15 +1,32 @@
 import { View, Text, StyleSheet } from "react-native";
+import MealsList from "../components/MealsList/MealsList";
+import { useContext } from "react";
+import { FavoritesContext } from "../store/context/favorites-context";
+import { MEALS } from "../data/dummy-data";
 
 function FavoritesScreen() {
+  const favoriteMealsCtx = useContext(FavoritesContext);
+
+  const favoriteMeals = MEALS.filter((meal) =>
+    favoriteMealsCtx.ids.includes(meal.id)
+  );
+
+  if (favoriteMeals.length === 0) {
+    return (
+      <View style={styles.rootContainer}>
+        <Text style={styles.text}>You have no favorite meals yet!</Text>
+      </View>
+    );
+  }
   return (
-    <View style={styles.container}>
-      <Text style={styles.text}>Favorites</Text>
+    <View style={styles.rootContainer}>
+      <MealsList items={favoriteMeals} />
     </View>
   );
 }
 
 const styles = StyleSheet.create({
-  container: {
+  rootContainer: {
     flex: 1,
     backgroundColor: "#3f2f25",
     alignItems: "center",
